+++ ee/tabby-webserver/src/schema/mod.rs
@@ -426,7 +426,7 @@ impl Mutation {
         avatar: Option<String>,
     ) -> Result<bool> {
         let claims = check_claims(ctx)?;
-        if claims.sub.0 != id && check_admin(ctx).is_err() {
+        if claims.sub.0 != id && check_admin(ctx).await.is_err() {
             return Err(CoreError::Unauthorized(
                 "You cannot change another user's avatar",
             ));
